Split tests into diff files
