+++ tests/integration/main.rs
@@ -1,4 +1,5 @@
 mod common;
 mod os;
+mod declare_txn_tests;
 mod pie;
 mod sharp;
